refactor(mcp): move generic ProseMirror primitives to lib/prosemirror/
These three modules have no notebook-specific knowledge — they're pure
functions over ProseMirror docs. Moving them out of `tools/notebooks/`
to `lib/prosemirror/` matches codebase convention (shared infra in
`lib/`, tool-specific code in `tools/<category>/`) and signals that
any future MCP tool editing PM content (comments, dashboards with rich
text, etc.) can reuse them.

Renames + clean up of notebook-specific wording:
  services/mcp/src/tools/notebooks/schema.ts    → lib/prosemirror/schema.ts
  services/mcp/src/tools/notebooks/diffSteps.ts → lib/prosemirror/diff.ts
  services/mcp/src/tools/notebooks/rebase.ts    → lib/prosemirror/rebase.ts
  services/mcp/tests/unit/notebooks-internals.test.ts → tests/unit/prosemirror.test.ts

Error messages in these modules are now generic ("refetch the source
document") — notebook-specific recovery hints stay where they belong,
in `tools/notebooks/edit.ts` which already owns the agent-facing
`not_found`/`no_content` messages with concrete tool names.

`saveLoop.ts` stays in `tools/notebooks/` since it knows the
collab/save URL and Notebook schema type — it's notebook-specific by
construction.
